+++ lib/hexapdf/layout/list_box.rb
@@ -227,12 +227,12 @@ module HexaPDF
             remove_indent_from_frame_shape(shape) unless shape.polygons.empty?
           end
 
-          item_frame = Frame.new(item_frame_left, top - height, item_frame_width, height,
-                                 shape: shape, context: frame.context)
+          item_frame = frame.child_frame(item_frame_left, top - height, item_frame_width, height,
+                                         shape: shape, box: self)
 
           if index != 0 || !split_box? || @split_box == :show_first_marker
             box = item_marker_box(frame.document, index)
-            marker_frame = Frame.new(0, 0, content_indentation, height, context: frame.context)
+            marker_frame = frame.child_frame(0, 0, content_indentation, height, box: self)
             break unless box.fit(content_indentation, height, marker_frame)
             item_result.marker = box
             item_result.marker_pos_x = item_frame.x - content_indentation